Extract the per-step trainers and their primitives out of voting_iterations
Stage 2 of splitting voting_iterations.py, in two new modules:

- vtscore/eval/step_model.py — the trainer-agnostic contract both sides need:
  StepModel, the head sentinels (HEADS / PRODUCTION_HEAD / resolve_hidden_dim),
  the cost weights, and the two functions that turn a media into a training
  vector or a simulation set into scores.  Nothing here imports the harness, so
  the dependency stays one-directional.
- vtscore/eval/step_trainers.py — the pool scorers and the *_train_and_calibrate
  family (~670 lines).

The six symbols now shared across module boundaries drop their underscores
(_StepModel -> StepModel, _resolve_hidden_dim -> resolve_hidden_dim, and so on);
the trainers themselves stay private to the harness.

_style_train_and_calibrate is pinned by check-eval-app-sync.py, so its Mirror's
harness= path moves to step_trainers.py.  The gate then passes *without*
--update, which is the check the move is byte-equivalent: the digest ignores
comments and formatting, so a re-pin would have been needed only if real logic
had changed.  Two notes in that script and one in vtscore/docs/packages/
training.md that named the old module are corrected in the same motion.

voting_iterations.py is down from 5145 to 3869 lines.

Refs #3405
