+++ vtscore/eval/voting_iterations.py
@@ -50,16 +50,27 @@ from vtscore.eval.al_strategies import ALContext, select_next
 from vtscore.eval.autopilot_flow import SMART_WINDOW, AutopilotFlow, app_has_detector
 from vtscore.eval.fit_quality import fit_quality_row
 from vtscore.eval.startup_schedule import StartupState, parse_startup_schedule, round_cut
-from vtscore.eval.labels import evaluable_pool, media_is_positive, region_box_for_category
+from vtscore.eval.step_model import (
+    HEADS,
+    PRODUCTION_HEAD,
+    StepModel,
+    inclusion_weights,
+    score_sim_set_with_model,
+)
+from vtscore.eval.step_trainers import (
+    _build_eval_atlas,
+    _labelset_error_costs,
+    _score_pool,
+    _train_and_calibrate,
+)
+from vtscore.eval.labels import evaluable_pool, media_is_positive
 from vtscore.eval.score_dumps import maybe_dump_predictions
-from vtscore.eval.trainers import _cross_calibrated_threshold, _parse_trainer_spec
 from vtscore.eval.voting_columns import (
     FIT_QUALITY_STRIDE_DEFAULT,
     SKYLINE_COLUMNS,
     VOTING_COLUMNS,
 )
 from vtscore.training.blend_schedules import BlendContext
-from vtscore.training.mlp import LINEAR_HEAD, LINEAR_SVM_HEAD, _auto_hidden_dim, train_model
 from vtscore.training.thresholds import (
     ACQUISITION_INCLUSION_OFFSET,
     CUT_KIND_INTERIOR,
@@ -70,62 +81,11 @@ from vtscore.training.thresholds import (
     acquisition_inclusion,
     anchored_gmm_fit,
     calculate_safe_threshold,
-    calibration_folds,
-    classify_threshold_provenance,
-    compute_fold_orderings,
-    compute_grouped_fold_node_scores,
     fold_anchored_gmm_threshold,
     gmm_cut_from_fit,
     rank_transfer,
     threshold_from_fold_orderings,
-    threshold_from_folds,
 )
-
-
-@dataclass
-class _StepModel:
-    """A trained per-step ranker plus the metadata the eval loop records.
-
-    ``predict`` maps an ``(N, D)`` numpy embedding matrix to per-row
-    ``P(positive)`` scores in ``[0, 1]`` — the trainer-agnostic scoring contract
-    (identical to :data:`vtscore.eval.trainers.PredictFn`).  ``torch_model`` is
-    set only for the MLP path, where region-aware datasets need the raw module
-    to max-pool over patch regions; it is ``None`` for the SVM path (which the
-    experiment only ever runs on single-vector, region-free datasets).
-    ``backend``/``device`` are recorded on every result row so the report can
-    say which engine produced each number.
-    """
-
-    predict: Callable[[Any], "np.ndarray"]
-    torch_model: Optional[Any]
-    backend: str
-    device: str
-
-
-#: Head choices for the harness's per-step ranker, all three reached through
-#: the same ``hidden_dim`` sentinel production threads.  ``"linear_svm"`` is the
-#: head the live detector trains (:data:`~vtscore.training.mlp.LINEAR_SVM_HEAD`,
-#: a single ``Linear(d, 1)`` fitted to the maximum-margin boundary), so a
-#: ``"linear_svm"`` run measures the shipped detector.  ``"linear"`` is the same
-#: architecture fitted with balanced BCE — logistic regression, the head shipped
-#: between #2790/#2809 and the SVM switch — and ``"mlp"`` is the older harness
-#: candidate, a hidden layer auto-sized from the vote count
-#: (:func:`~vtscore.training.mlp._auto_hidden_dim`).  The choice is threaded into
-#: the calibration folds too, exactly as production threads one sentinel through
-#: ``_train_and_score_xy``.
-HEADS: tuple[str, ...] = ("mlp", "linear", "linear_svm")
-
-#: The head the **app** trains, and therefore the harness's default arm:
-#: ``vtscore.detectors.training.train_and_threshold`` pins ``hidden_dim =
-#: LINEAR_SVM_HEAD`` on every production fit.  ``head=None`` resolves to this,
-#: the way ``style=None`` and ``blend_schedule=None`` resolve to the app's
-#: geometry and blend schedule — an eval default that isn't the app default
-#: measures a detector nobody ships (see the "Eval Default Arm IS the App"
-#: rule).  If the shipped head ever changes, move this with it:
-#: ``test_harness_linear_head`` pins the two against each other by training the
-#: real app pipeline, so the suite fails rather than letting the default arm
-#: drift silently.
-PRODUCTION_HEAD: str = "linear_svm"
 
 
 #: The detection geometry a live detector uses on a patch dataset - the style an
@@ -184,21 +144,6 @@ SKYLINE_PROVENANCE: str = "skyline_test_oracle"
 _SKYLINE_SEED: int = 3322
 
 
-def _resolve_hidden_dim(head: str, n_votes: int) -> int:
-    """``hidden_dim`` sentinel for *head* at *n_votes* votes.
-
-    The two linear heads return their sentinels (they have no width to size);
-    only ``"mlp"`` consults the vote count.
-    """
-    if head == "linear_svm":
-        return LINEAR_SVM_HEAD
-    if head == "linear":
-        return LINEAR_HEAD
-    if head == "mlp":
-        return _auto_hidden_dim(n_votes)
-    raise ValueError(f"unknown head {head!r}; expected one of {HEADS}")
-
-
 # ------------------------------------------------------------------
 # Helpers
 # ------------------------------------------------------------------
@@ -208,17 +153,6 @@ def _resolve_hidden_dim(head: str, n_votes: int) -> int:
 #: this the held-out test set has too few positives for a stable FNR estimate,
 #: so the arm is skipped rather than reported with a noisy denominator.
 _MIN_PREVALENCE_POSITIVES = 15
-
-
-def _inclusion_weights(inclusion: int) -> tuple[float, float]:
-    """``(fpr_weight, fnr_weight)`` for an inclusion value.
-
-    Delegates to the production definition so a measured cost and the shipped
-    threshold rule can never disagree about what an inclusion value prices.
-    """
-    from vtscore.training.thresholds import inclusion_cost_weights  # noqa: PLC0415
-
-    return inclusion_cost_weights(inclusion)
 
 
 def _prevalence(clips_dict: dict[int, dict[str, Any]], target_category: str) -> float:
@@ -269,67 +203,6 @@ def _split_media_ids(
     shuffled = rng.permutation(all_ids).tolist()
     n_sim = max(1, int(len(shuffled) * sim_fraction))
     return shuffled[:n_sim], shuffled[n_sim:]
-
-
-def _good_training_vec(
-    media: dict[str, Any],
-    target_category: str,
-    region_voting: bool,
-) -> np.ndarray:
-    """Return the training vector for one Good vote on *media*.
-
-    With *region_voting* the simulated user drags the ground-truth box around
-    the object: when *media* carries a stored ``patch_grid`` and an annotated
-    region for *target_category*, the box is pooled on-the-fly via
-    :func:`vtscore.detectors.training.pool_box_from_media` (the same path the
-    live region-vote flow uses).  Falls back to the whole-image embedding when
-    region voting is off, the media has no patch grid (single-vector
-    embedders), or no box is annotated for this category - exactly an
-    image-level Good vote.
-    """
-    if region_voting:
-        from vtscore.detectors.training import pool_box_from_media  # noqa: PLC0415
-
-        pooled = pool_box_from_media(media, region_box_for_category(media, target_category))
-        if pooled is not None:
-            return pooled
-    return media_embedding(media)
-
-
-def _score_sim_set_with_model(
-    model: Any,
-    region_aware: bool,
-    sim_clips: dict[int, dict[str, Any]] | None,
-    X_all_clips: Any,
-    sim_ids: list[int],
-    style_obj: Any = None,
-) -> tuple[list[int], list[float]]:
-    """``(ids, scores)`` for the simulation set under an arbitrary torch *model*.
-
-    The same scorer the test set uses, so the population estimator sees the
-    distribution the threshold will actually cut: through the detection style
-    when one is given, else region max-pool on a patch dataset, else the
-    pre-computed whole-image matrix *X_all_clips* (stacked over
-    ``sorted(sim_ids)`` - that ordering is preserved).
-    """
-    import numpy as np  # noqa: PLC0415
-    import torch  # noqa: PLC0415
-
-    if style_obj is not None:
-        assert sim_clips is not None
-        score_map = style_obj.score_media(model, sim_clips)
-        ids = list(score_map.keys())
-        return ids, [float(score_map[cid]) for cid in ids]
-    if region_aware:
-        from vtscore.detectors.training import score_media_with_model  # noqa: PLC0415
-
-        assert sim_clips is not None
-        scored = score_media_with_model(model, sim_clips)
-        return [int(r["id"]) for r in scored], [float(r["score"]) for r in scored]
-    with torch.no_grad():
-        t = torch.tensor(np.asarray(X_all_clips), dtype=torch.float32).to(next(model.parameters()).device)
-        scores = torch.sigmoid(model(t)).squeeze(1).cpu().numpy()
-    return sorted(sim_ids), [float(s) for s in scores]
 
 
 def _pool_percentile(pool_scores: dict[int, float], threshold: float) -> float:
@@ -395,7 +268,7 @@ def _blend_xcal_input(threshold: float, details: dict[str, Any]) -> float:
 
 def _safe_threshold_for_step(
     threshold: float,
-    step: _StepModel,
+    step: StepModel,
     details: dict[str, Any],
     region_aware: bool,
     sim_clips: dict[int, dict[str, Any]] | None,
@@ -471,7 +344,7 @@ def _safe_threshold_for_step(
     t_final = time.monotonic()
     if style_obj is not None or region_aware:
         assert final_model is not None
-        ids, all_scores = _score_sim_set_with_model(
+        ids, all_scores = score_sim_set_with_model(
             final_model, region_aware, sim_clips, X_all_clips, sim_ids, style_obj
         )
     else:
@@ -506,7 +379,7 @@ def _safe_threshold_for_step(
     haystack_seconds: list[float] = []
     for model in fold_models[:n_folds]:
         t_hay = time.monotonic()
-        fids, fscores = _score_sim_set_with_model(model, region_aware, sim_clips, X_all_clips, sim_ids, style_obj)
+        fids, fscores = score_sim_set_with_model(model, region_aware, sim_clips, X_all_clips, sim_ids, style_obj)
         hay = _hay(fscores, fids)
         haystack_seconds.append(time.monotonic() - t_hay)
         fold_haystacks.append(hay)
@@ -524,7 +397,7 @@ def _safe_threshold_for_step(
         extended = list(fold_haystacks)
         for model in fold_data["models"][len(extended) :]:
             t_hay = time.monotonic()
-            fids, fscores = _score_sim_set_with_model(model, region_aware, sim_clips, X_all_clips, sim_ids, style_obj)
+            fids, fscores = score_sim_set_with_model(model, region_aware, sim_clips, X_all_clips, sim_ids, style_obj)
             hay = _hay(fscores, fids)
             haystack_seconds.append(time.monotonic() - t_hay)
             extended.append(hay)
@@ -546,7 +419,7 @@ def _safe_threshold_for_step(
 
 
 def _evaluate_on_test(
-    step: _StepModel,
+    step: StepModel,
     threshold: float,
     clips_dict: dict[int, dict[str, Any]],
     test_ids: list[int],
@@ -627,7 +500,7 @@ def _evaluate_on_test(
 
     maybe_dump_predictions(clips_dict, test_ids, scores, true_labels, threshold, target_category, suffix="__eval")
 
-    fpr_weight, fnr_weight = _inclusion_weights(inclusion)
+    fpr_weight, fnr_weight = inclusion_weights(inclusion)
     cost = fpr_weight * fpr + fnr_weight * fnr
 
     scores_arr = np.asarray(scores, dtype=np.float64)
@@ -1939,7 +1812,7 @@ def _emit_fold_anchored_rows(
 
 
 def _calibration_metric_rows(
-    step: _StepModel,
+    step: StepModel,
     threshold: float,
     details: dict[str, Any],
     clips_dict: dict[int, dict[str, Any]],
@@ -2251,683 +2124,6 @@ def _cut_inclusion_arms(
 
 
 # ------------------------------------------------------------------
-# Active-learning acquisition helpers
-# ------------------------------------------------------------------
-
-
-def _score_pool(
-    step: _StepModel,
-    pool_ids: list[int],
-    clips_dict: dict[int, dict[str, Any]],
-    *,
-    region_aware: bool = False,
-    style_obj: Any = None,
-    sim_clips: dict[int, dict[str, Any]] | None = None,
-    sim_scored: tuple[list[int], list[float]] | None = None,
-) -> dict[int, float]:
-    """Return ``{pool_id: score}`` for the current model over the pool.
-
-    **In the same score space the thresholds are cut in** (issue #2943).  That
-    is not a refinement, it is a correctness requirement: the Hard pick locates
-    its cutoff with the *absolute* comparison ``ranking[cid] <= threshold``
-    (:func:`~vtscore.eval.al_strategies._hard_pick_by_index`), so a ranking and
-    a cut that live in different spaces put the cutoff index in the wrong place.
-    On a patch dataset the reporting/acquisition cuts are fitted on the style's
-    region max-pooled scores, and a max over ~197 patch rows stochastically
-    dominates the single whole-image row - so scoring the pool whole-image would
-    depress every pool score relative to the cut and drag the cutoff index
-    systematically toward the top of the ranking.  The app has no such gap: its
-    learned sort ranks the very same pooled scores its threshold cuts.
-
-    Three paths, mirroring :func:`_evaluate_on_test` / :func:`_score_sim_set_with_model`:
-
-    * *sim_scored* - the ``(ids, scores)`` the safe-threshold step already
-      computed over the whole simulation set, in exactly this geometry.  The
-      pool is a subset of that set, so restricting it is free and removes the
-      scoring pass entirely.
-    * a *style_obj* / *region_aware* dataset with no such scores (the
-      ``safe_thresholds=False`` control arm) - score through the style.  The
-      **full** sim set is scored rather than just the pool: the style memoises
-      its flattened patch matrix per media-id set, and the pool loses an item
-      every step, so scoring the shrinking pool would re-flatten from scratch
-      each step *and* leak a cache entry per step.
-    * everything else (single-vector datasets, the SVM arms) - the trainer-
-      agnostic whole-image ``predict``, which is already the threshold's space.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    if not pool_ids:
-        return {}
-    if sim_scored is not None:
-        ids, scores = sim_scored
-        pool_set = set(pool_ids)
-        return {cid: float(s) for cid, s in zip(ids, scores, strict=True) if cid in pool_set}
-    if (style_obj is not None or region_aware) and sim_clips:
-        assert step.torch_model is not None
-        pool_set = set(pool_ids)
-        ids, scores = _score_sim_set_with_model(
-            step.torch_model, region_aware, sim_clips, None, sorted(sim_clips), style_obj
-        )
-        return {cid: float(s) for cid, s in zip(ids, scores, strict=True) if cid in pool_set}
-    embs = np.array([media_embedding(clips_dict[cid]) for cid in pool_ids])
-    scores = np.asarray(step.predict(embs)).ravel().tolist()
-    return dict(zip(pool_ids, scores, strict=True))
-
-
-def _labelset_error_costs(
-    model_steps: list[tuple[Any, float]],
-    good_votes: dict[int, None],
-    bad_votes: dict[int, None],
-    clips_dict: dict[int, dict[str, Any]],
-    inclusion: int,
-) -> list[float]:
-    """Weighted FPR/FNR of **every** recent model on the current labelled set.
-
-    Feeds the Smart indicator.  Mirrors ``labeling_progress._eval_cached_models``
-    /``_score_step``: every model in the window is re-scored against the
-    *current* labelset — the only ground truth the app has — with its own cached
-    threshold, so all points of the slope regression share one eval set and the
-    trend isolates model improvement.  Scoring each model against the labelset
-    it was trained on instead would confound model change with labelset growth:
-    autopilot deliberately votes boundary items, which are mispredicted at first
-    and inflate the later costs of a frozen-cost history.
-
-    Deliberately *not* the held-out test split: those labels must never reach
-    the vote order.  Returns ``[]`` when the labelset has no usable eval set
-    (either class empty), matching ``_eval_cached_models``, which leaves the
-    Smart indicator on its "not enough points" branch.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    ids = list(good_votes) + list(bad_votes)
-    labels = [1.0] * len(good_votes) + [0.0] * len(bad_votes)
-    total_pos = len(good_votes)
-    total_neg = len(bad_votes)
-    if not model_steps or not ids or total_pos == 0 or total_neg == 0:
-        return []
-
-    fpr_weight, fnr_weight = _inclusion_weights(inclusion)
-    # One eval matrix, reused by every model in the window - the app's
-    # ``_build_eval_set`` builds its tensor once for the same reason.
-    embs = np.array([media_embedding(clips_dict[cid]) for cid in ids])
-
-    costs: list[float] = []
-    for step, threshold in model_steps:
-        scores = np.asarray(step.predict(embs)).ravel()
-        fp = fn = 0
-        for score, true_label in zip(scores.tolist(), labels, strict=True):
-            predicted = 1 if score >= threshold else 0
-            if predicted == 1 and true_label == 0.0:
-                fp += 1
-            elif predicted == 0 and true_label == 1.0:
-                fn += 1
-        fpr = fp / total_neg
-        fnr = fn / total_pos
-        costs.append(fpr_weight * fpr + fnr_weight * fnr)
-    return costs
-
-
-def _build_eval_atlas(embeddings: dict[int, np.ndarray], min_node_size: int) -> Any:
-    """Build a coverage atlas over *embeddings* for the autopilot New phase.
-
-    Returns ``None`` when there are no vectors.  Uses the same hierarchical
-    k-means partition the live dataset builds (see
-    :class:`~vtscore.coverage.atlas.CoverageAtlas`); *min_node_size* is
-    exposed so a caller with a small simulation set can drive the partition
-    deeper than the production floor (20) and actually resolve density cells.
-    """
-    from vtscore.coverage.atlas import CoverageAtlas, auto_max_depth  # noqa: PLC0415
-
-    if not embeddings:
-        return None
-    return CoverageAtlas(
-        embeddings,
-        k=3,
-        max_depth=auto_max_depth(len(embeddings), k=3, min_node_size=min_node_size),
-        min_node_size=min_node_size,
-    )
-
-
-def _train_and_calibrate(
-    trainer: str,
-    good_votes: dict[int, None],
-    bad_votes: dict[int, None],
-    clips_dict: dict[int, dict[str, Any]],
-    target_category: str,
-    *,
-    region_voting: bool,
-    input_dim: int,
-    inclusion: int,
-    calibrate_count: int,
-    calibration_fraction: float,
-    head: str = PRODUCTION_HEAD,
-    style_obj: Any = None,
-    emit_calibration_metrics: bool = False,
-    fold_count_variants: list[int] | None = None,
-) -> tuple[_StepModel, float, int, dict[str, float], dict[str, Any]]:
-    """Train the step's ranker and calibrate its threshold from the current votes.
-
-    *head* selects the head on both production paths (see :data:`HEADS`):
-    ``"linear_svm"`` (the default, :data:`PRODUCTION_HEAD`) trains the head the
-    live detector has, ``"linear"`` the logistic head it replaced, ``"mlp"`` the
-    legacy auto-sized hidden layer.  It is ignored by the standalone SVM path,
-    which fits its own estimator rather than a head.
-
-    Dispatches on *trainer*: ``"mlp"`` runs the production MLP path unchanged
-    (see :func:`_mlp_train_and_calibrate`); any ``svm_*`` name runs the SVM path
-    (see :func:`_svm_train_and_calibrate`).  Returns ``(step, threshold,
-    n_labels, timings, details)`` where *timings* has ``train_seconds`` and
-    ``xcal_seconds`` for the fit and threshold-calibration wall clocks, and
-    *details* is empty unless *emit_calibration_metrics* (the #2781 study),
-    carrying the fold orderings, node scores, and threshold provenance the
-    calibration metrics need.
-
-    With an explicit *style_obj* (MLP only) the vote-to-vector assembly is
-    delegated to the style (see :func:`_style_train_and_calibrate`).
-    """
-    if style_obj is not None:
-        return _style_train_and_calibrate(
-            style_obj,
-            good_votes,
-            bad_votes,
-            clips_dict,
-            target_category,
-            region_voting=region_voting,
-            input_dim=input_dim,
-            inclusion=inclusion,
-            calibrate_count=calibrate_count,
-            calibration_fraction=calibration_fraction,
-            head=head,
-            emit_calibration_metrics=emit_calibration_metrics,
-            fold_count_variants=fold_count_variants,
-        )
-    if trainer == "mlp":
-        return _mlp_train_and_calibrate(
-            good_votes,
-            bad_votes,
-            clips_dict,
-            target_category,
-            region_voting=region_voting,
-            input_dim=input_dim,
-            inclusion=inclusion,
-            calibrate_count=calibrate_count,
-            calibration_fraction=calibration_fraction,
-            head=head,
-        )
-    return _svm_train_and_calibrate(
-        trainer,
-        good_votes,
-        bad_votes,
-        clips_dict,
-        target_category,
-        inclusion=inclusion,
-        calibrate_count=calibrate_count,
-        calibration_fraction=calibration_fraction,
-    )
-
-
-def _mlp_train_and_calibrate(
-    good_votes: dict[int, None],
-    bad_votes: dict[int, None],
-    clips_dict: dict[int, dict[str, Any]],
-    target_category: str,
-    *,
-    region_voting: bool,
-    input_dim: int,
-    inclusion: int,
-    calibrate_count: int,
-    calibration_fraction: float,
-    head: str = PRODUCTION_HEAD,
-) -> tuple[_StepModel, float, int, dict[str, float], dict[str, Any]]:
-    """The production arm — numerically identical to the pre-trainer harness at ``head="mlp"``.
-
-    At ``head="linear_svm"`` (the default, :data:`PRODUCTION_HEAD`) this trains
-    the live detector's head: production pins the linear SVM on every fit (see
-    ``vtscore.training.mlp.LINEAR_SVM_HEAD``), so the reported thresholds and
-    costs are the shipped detector's.  ``head="linear"`` (the logistic head the
-    SVM replaced) and ``head="mlp"`` (the small-MLP candidate #2781 measured)
-    are the named legacy arms.  Everything *around* the head mirrors the
-    production ``_train_and_score_xy`` / ``train_and_threshold`` pipeline
-    whichever is chosen:
-
-    Good votes region-pool their ground-truth box when *region_voting* is on
-    (and the media supports it); Bad votes always train on the whole-image
-    vector.
-
-    **This is the single-vector path.**  Bad votes here are one row because a
-    single-vector media *has* one row - not because the live detector works that
-    way.  On a patch dataset the live detector floods a Bad vote over the
-    image's whole score-row stack, and
-    :func:`simulate_voting_iterations` routes such datasets to the
-    ``max_patch`` style (:func:`_style_train_and_calibrate`) rather than here,
-    so the default arm matches the app.  Do not "restore" whole-image Bad votes
-    on patch data: that trains ~196 rows per rejected image down never while
-    inference max-pools them.
-
-    * ``hidden_dim`` comes from the head (sized from the *full* label count on
-      the MLP head, 0 on the linear one) and is forced onto the
-      calibration folds, so the fold models share the final model's architecture
-      (production likewise threads one width into
-      ``cross_calibration_threshold_cached``).  Letting each fold auto-size to
-      its own smaller train split would train narrower fold nets and report a
-      threshold no single-architecture pipeline ever produces.
-    * the fold splits use a fresh ``RandomState(42)`` - the fixed seed
-      ``cross_calibration_threshold_cached`` always calibrates with - rather than
-      the shared per-seed simulation RNG, so the calibration is byte-for-byte
-      what production runs for this vote set.  The eval seed still varies the
-      data (which media are voted, in what order, and the held-out test split);
-      only the calibration folds are pinned, as they are in production.
-    """
-    import numpy as np  # noqa: PLC0415
-    import torch  # noqa: PLC0415
-
-    X_list: list[np.ndarray] = []
-    y_list: list[float] = []
-    for vid in good_votes:
-        X_list.append(_good_training_vec(clips_dict[vid], target_category, region_voting))
-        y_list.append(1.0)
-    for vid in bad_votes:
-        X_list.append(media_embedding(clips_dict[vid]))
-        y_list.append(0.0)
-
-    X = torch.tensor(np.array(X_list), dtype=torch.float32)
-    y = torch.tensor(y_list, dtype=torch.float32).unsqueeze(1)
-    n_labels = len(good_votes) + len(bad_votes)
-
-    hidden_dim = _resolve_hidden_dim(head, n_labels)
-    t_xcal = time.monotonic()
-    # The folds' orderings *and* models ride out in ``details`` unconditionally:
-    # the shipped safe threshold anchors on the fold models' held-out scores, so
-    # they are an input to the baseline arm, not study-only extras.
-    folds = calibration_folds(
-        X_list,
-        y_list,
-        input_dim,
-        calibrate_count=calibrate_count,
-        calibration_fraction=calibration_fraction,
-        hidden_dim=hidden_dim,
-        rng=np.random.RandomState(42),
-    )
-    threshold = threshold_from_folds(folds, inclusion)
-    xcal_seconds = time.monotonic() - t_xcal
-    t_train = time.monotonic()
-    model = train_model(X, y, input_dim, hidden_dim=hidden_dim)
-    train_seconds = time.monotonic() - t_train
-
-    device = str(next(model.parameters()).device)
-
-    def predict(X_test: Any) -> np.ndarray:
-        with torch.no_grad():
-            t = torch.tensor(np.asarray(X_test), dtype=torch.float32).to(next(model.parameters()).device)
-            return torch.sigmoid(model(t)).squeeze(1).cpu().numpy()
-
-    step = _StepModel(
-        predict=predict,
-        torch_model=model,
-        backend="torch-cuda" if device.startswith("cuda") else "torch-cpu",
-        device=device,
-    )
-    details = {
-        "fold_orderings": folds.orderings,
-        "fold_models": folds.models,
-        # Which sentinel (if any) the fold rule returned: the blend's x-cal side
-        # is NO_GOOD_THRESHOLD whenever this is set, as production's does
-        # (see :func:`_blend_xcal_input`).
-        "fold_fallback": folds.fallback,
-    }
-    return step, threshold, n_labels, {"train_seconds": train_seconds, "xcal_seconds": xcal_seconds}, details
-
-
-def _style_train_and_calibrate(
-    style_obj: Any,
-    good_votes: dict[int, None],
-    bad_votes: dict[int, None],
-    clips_dict: dict[int, dict[str, Any]],
-    target_category: str,
-    *,
-    region_voting: bool,
-    input_dim: int,
-    inclusion: int,
-    calibrate_count: int,
-    calibration_fraction: float,
-    head: str = PRODUCTION_HEAD,
-    emit_calibration_metrics: bool = False,
-    fold_count_variants: list[int] | None = None,
-) -> tuple[_StepModel, float, int, dict[str, float], dict[str, Any]]:
-    """Style-driven torch path (the Max-Patch experiment arms).
-
-    The detection style (see :mod:`vtscore.eval.patch_styles`) supplies the
-    vote-to-vector rules: each Good vote contributes ``style.good_vec`` (given
-    the ground-truth box when *region_voting* and the media has one), each Bad
-    vote floods ``style.bad_vecs`` - one row on a whole-image style, the
-    image-level vector + every raw patch on ``max_patch``, every tree node on
-    the HAC hybrids.
-
-    Training and calibration are **bag-aware**, exactly like the production
-    vote path (:func:`vtscore.detectors.training._train_and_score_xy`): the
-    head (see :data:`HEADS`) and the safe-threshold ramp size on distinct *votes* rather
-    than flooded rows, the calibration folds split by bag, and the final fit
-    weights each bag equally.  On a whole-image style every bag is one row, so
-    this collapses to the historical single-vector behaviour.
-
-    Calibration additionally runs in **inference geometry**: each bag is handed
-    its ``style.score_rows`` stack so a Good bag collapses the same way a Bad
-    bag (and every held-out image) does.  Without this a Good bag is a max over
-    its 1 training row while a Bad bag is a max over the ~197 rows it flooded,
-    and the calibrated cut lands above the score range production actually
-    produces - see :func:`vtscore.training.thresholds.compute_fold_orderings`.
-    """
-    import numpy as np  # noqa: PLC0415
-    import torch  # noqa: PLC0415
-
-    from vtscore.detectors.training import _flood_context  # noqa: PLC0415
-
-    X_list: list[np.ndarray] = []
-    y_list: list[float] = []
-    groups: list = []
-    score_rows_by_group: dict = {}
-    for vid in good_votes:
-        box = region_box_for_category(clips_dict[vid], target_category) if region_voting else None
-        X_list.append(np.asarray(style_obj.good_vec(clips_dict[vid], box), dtype=np.float32))
-        y_list.append(1.0)
-        groups.append(("g", vid))
-        score_rows_by_group[("g", vid)] = style_obj.score_rows(clips_dict[vid])
-    for vid in bad_votes:
-        for vec in style_obj.bad_vecs(clips_dict[vid]):
-            X_list.append(np.asarray(vec, dtype=np.float32))
-            y_list.append(0.0)
-            groups.append(("b", vid))
-        score_rows_by_group[("b", vid)] = style_obj.score_rows(clips_dict[vid])
-
-    X = torch.tensor(np.array(X_list), dtype=torch.float32)
-    y = torch.tensor(y_list, dtype=torch.float32).unsqueeze(1)
-    n_votes, cal_groups, sample_weights = _flood_context(X_list, y_list, groups)
-
-    hidden_dim = _resolve_hidden_dim(head, n_votes)
-    t_xcal = time.monotonic()
-    details: dict[str, Any] = {}
-    if emit_calibration_metrics:
-        threshold, details = _calibrate_with_details(
-            X_list,
-            y_list,
-            input_dim,
-            inclusion,
-            calibrate_count=calibrate_count,
-            calibration_fraction=calibration_fraction,
-            hidden_dim=hidden_dim,
-            cal_groups=cal_groups,
-            score_rows_by_group=score_rows_by_group if cal_groups is not None else None,
-            fold_count_variants=fold_count_variants,
-        )
-        # Bad-voted bags' inference row stacks: the final model scores these to
-        # form the pnorm null (F_neg) at test time (see _calibration_metric_rows).
-        details["neg_score_rows"] = [score_rows_by_group[("b", vid)] for vid in bad_votes]
-    else:
-        # Same fold work as the metrics branch, minus the study extras: the
-        # shipped safe threshold anchors on the fold models, so they ride out
-        # in ``details`` on every path (see :func:`_safe_threshold_for_step`).
-        folds = calibration_folds(
-            X_list,
-            y_list,
-            input_dim,
-            calibrate_count=calibrate_count,
-            calibration_fraction=calibration_fraction,
-            hidden_dim=hidden_dim,
-            rng=np.random.RandomState(42),
-            groups=cal_groups,
-            score_rows_by_group=score_rows_by_group if cal_groups is not None else None,
-        )
-        threshold = threshold_from_folds(folds, inclusion)
-        details = {
-            "fold_orderings": folds.orderings,
-            "fold_models": folds.models,
-            "fold_fallback": folds.fallback,
-        }
-    xcal_seconds = time.monotonic() - t_xcal
-    # Under the #2897 screen this step trained Kmax folds, not ``calibrate_count``
-    # of them.  Bill the reported wall clock for the live count only, so the
-    # baseline row's timing stays the one an uninstrumented run would report; the
-    # per-K costs live in each fold-count arm's own ``fold_seconds``.
-    extra = (details.get("fold_count_data") or {}).get("seconds")
-    if extra:
-        xcal_seconds -= sum(extra[calibrate_count:])
-    t_train = time.monotonic()
-    if sample_weights is not None:
-        model = train_model(X, y, input_dim, hidden_dim=hidden_dim, sample_weights=sample_weights)
-    else:
-        model = train_model(X, y, input_dim, hidden_dim=hidden_dim)
-    train_seconds = time.monotonic() - t_train
-
-    device = str(next(model.parameters()).device)
-
-    def predict(X_test: Any) -> np.ndarray:
-        with torch.no_grad():
-            t = torch.tensor(np.asarray(X_test), dtype=torch.float32).to(next(model.parameters()).device)
-            return torch.sigmoid(model(t)).squeeze(1).cpu().numpy()
-
-    step = _StepModel(
-        predict=predict,
-        torch_model=model,
-        backend="torch-cuda" if device.startswith("cuda") else "torch-cpu",
-        device=device,
-    )
-    return step, threshold, n_votes, {"train_seconds": train_seconds, "xcal_seconds": xcal_seconds}, details
-
-
-def _calibrate_with_details(
-    X_list: list[np.ndarray],
-    y_list: list[float],
-    input_dim: int,
-    inclusion: int,
-    *,
-    calibrate_count: int,
-    calibration_fraction: float,
-    hidden_dim: int | None,
-    cal_groups: list | None,
-    score_rows_by_group: dict | None,
-    fold_count_variants: list[int] | None = None,
-) -> tuple[float, dict[str, Any]]:
-    """Compute the trained threshold **and** the calibration study's provenance.
-
-    Replaces the plain :func:`calculate_cross_calibration_threshold` call on the
-    style path when the #2781 metrics are requested.  Trains the calibration
-    folds exactly once and returns ``(threshold, details)`` where *details* holds:
-
-    * ``provenance`` — which code path set the threshold (``conformal`` /
-      ``no_good_sentinel`` / ``too_few_default``), via
-      :func:`~vtscore.training.thresholds.classify_threshold_provenance`.
-    * ``fold_orderings`` — the pooled ``(scores, labels)`` per fold under the
-      base (max) pooling, for the calibration-set oracle and the inclusion sweep.
-    * ``fold_node_data`` — per-fold, per-group **node** scores (grouped path
-      only), so a remedial pooling variant can recalibrate off the same fold
-      models without retraining; ``None`` on the row-wise (whole-image) path.
-    * ``fold_fallback`` — the sentinel the fold rule returned, or ``None`` when
-      the folds are real.  The shipped blend substitutes ``NO_GOOD_THRESHOLD``
-      for the x-cal side whenever this is set, as production does (see
-      :func:`_blend_xcal_input`).
-    * ``fold_count_data`` — only under *fold_count_variants* (issue #2897): the
-      **full** Kmax fold orderings, their per-fold seconds, and the
-      count-independent overhead, for :func:`_fold_count_variant_rows`.
-
-    On the grouped path the fold models are trained once via
-    :func:`~vtscore.training.thresholds.compute_grouped_fold_node_scores` and the
-    base orderings are the max-pool of the node data, so the threshold is
-    identical to what production's grouped calibration produces for this arm.
-
-    *fold_count_variants* raises the number of folds actually trained to
-    ``max(calibrate_count, *variants)`` while leaving everything the step
-    returns computed off the first ``calibrate_count`` of them.  That is exact,
-    not an approximation: the folds are nested (see
-    :func:`~vtscore.training.thresholds.compute_fold_orderings`) and
-    ``train_model`` is seeded per call, so the extra folds cannot perturb the
-    live threshold, the fold models, or the trajectory - they only cost time.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    k_max = max(calibrate_count, *(fold_count_variants or [calibrate_count]))
-    t_folds = time.monotonic()
-    fold_seconds: list[float] = []
-
-    def _with_fold_data(details: dict[str, Any], orderings: list) -> dict[str, Any]:
-        """Attach the fold-count screen's inputs and trim *details* to K live folds."""
-        if fold_count_variants:
-            details["fold_count_data"] = {
-                "orderings": orderings,
-                # The **untrimmed** fold models, so the #3116 anchored arm can
-                # re-fit production's rule at every K.  `details["fold_models"]`
-                # is deliberately cut to the live count so nothing downstream
-                # can accidentally widen the shipped threshold's own fit.
-                "models": list(fold_models),
-                "seconds": fold_seconds,
-                # Everything in the calibration wall clock that is *not* a fold
-                # fit (the pooled conformal rule, the node max-pool): paid once
-                # at every K, so it belongs in each arm's cost.
-                "overhead_seconds": max(0.0, (time.monotonic() - t_folds) - sum(fold_seconds)),
-            }
-        return details
-
-    # The trained fold models ride along in details["fold_models"] so the
-    # #2852 fold-anchored arm can score the haystack on each fold's own scale
-    # without retraining; production callers never see them.
-    fold_models: list = []
-    if cal_groups is not None:
-        fold_node_data, fallback = compute_grouped_fold_node_scores(
-            X_list,
-            y_list,
-            input_dim,
-            groups=cal_groups,
-            rng=np.random.RandomState(42),
-            calibrate_count=k_max,
-            calibration_fraction=calibration_fraction,
-            hidden_dim=hidden_dim,
-            score_rows_by_group=score_rows_by_group,
-            model_sink=fold_models,
-            seconds_sink=fold_seconds,
-        )
-        if fallback is not None:
-            return fallback, {
-                "provenance": classify_threshold_provenance(fallback),
-                "fold_orderings": [],
-                "fold_node_data": None,
-                "fold_models": [],
-                "fold_fallback": fallback,
-            }
-        # Base (max) orderings from the same fold node data -> identical to
-        # production's grouped calibration for this arm.
-        all_orderings = [([float(np.max(b)) for b in blocks], labels) for blocks, labels in fold_node_data]
-        fold_orderings = all_orderings[:calibrate_count]
-        threshold = threshold_from_fold_orderings(fold_orderings, inclusion)
-        return threshold, _with_fold_data(
-            {
-                "provenance": classify_threshold_provenance(None),
-                "fold_orderings": fold_orderings,
-                "fold_node_data": fold_node_data[:calibrate_count],
-                "fold_models": fold_models[:calibrate_count],
-                "fold_fallback": None,
-            },
-            all_orderings,
-        )
-
-    # Row-wise path (whole-image styles): no bag flooding, no node re-pooling.
-    all_orderings, fallback = compute_fold_orderings(
-        X_list,
-        y_list,
-        input_dim,
-        rng=np.random.RandomState(42),
-        calibrate_count=k_max,
-        calibration_fraction=calibration_fraction,
-        hidden_dim=hidden_dim,
-        model_sink=fold_models,
-        seconds_sink=fold_seconds,
-    )
-    if fallback is not None:
-        return fallback, {
-            "provenance": classify_threshold_provenance(fallback),
-            "fold_orderings": [],
-            "fold_node_data": None,
-            "fold_models": [],
-            "fold_fallback": fallback,
-        }
-    fold_orderings = all_orderings[:calibrate_count]
-    threshold = threshold_from_fold_orderings(fold_orderings, inclusion)
-    return threshold, _with_fold_data(
-        {
-            "provenance": classify_threshold_provenance(None),
-            "fold_orderings": fold_orderings,
-            "fold_node_data": None,
-            "fold_models": fold_models[:calibrate_count],
-            "fold_fallback": None,
-        },
-        all_orderings,
-    )
-
-
-def _svm_train_and_calibrate(
-    trainer: str,
-    good_votes: dict[int, None],
-    bad_votes: dict[int, None],
-    clips_dict: dict[int, dict[str, Any]],
-    target_category: str,
-    *,
-    inclusion: int,
-    calibrate_count: int,
-    calibration_fraction: float,
-) -> tuple[_StepModel, float, int, dict[str, float], dict[str, Any]]:
-    """SVM path — single-vector only (the experiment never region-votes an SVM).
-
-    Threshold uses the trainer-agnostic cross-calibration port
-    (:func:`vtscore.eval.trainers._cross_calibrated_threshold`) — the natural
-    analogue of the MLP's production calibration — with the fold models pinned
-    to the sklearn CPU backend (they are tiny and only feed the threshold, so
-    paying GPU launch overhead per fold would be wasteful).  The *final* fit
-    honours the ambient backend (cuML on a GPU unless ``VTSEARCH_DISABLE_CUML``
-    forces sklearn), and that backend is what the row records and what produces
-    the scores.  The SVM fit seed is pinned to 42, mirroring the MLP's fixed
-    calibration seed; the eval seed still varies which items are voted.
-    """
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.eval.trainers import _train_svm_factory  # noqa: PLC0415
-    from vtscore.training.svm import train_svm  # noqa: PLC0415
-
-    X = np.array(
-        [media_embedding(clips_dict[vid]) for vid in good_votes]
-        + [media_embedding(clips_dict[vid]) for vid in bad_votes],
-        dtype=np.float32,
-    )
-    y = np.array([1] * len(good_votes) + [0] * len(bad_votes), dtype=np.int32)
-    n_labels = len(good_votes) + len(bad_votes)
-
-    kernel, kwargs = _parse_trainer_spec(trainer)
-
-    # Fold models for the threshold are pinned to sklearn CPU (tiny fits).
-    fold_trainer = _train_svm_factory(kernel, backend="sklearn", **kwargs)
-    t_xcal = time.monotonic()
-    threshold = _cross_calibrated_threshold(
-        X,
-        y,
-        fold_trainer,
-        42,
-        inclusion_value=inclusion,
-        calibrate_count=calibrate_count,
-        cal_fraction=calibration_fraction,
-    )
-    xcal_seconds = time.monotonic() - t_xcal
-
-    t_train = time.monotonic()
-    clf = train_svm(X, y, kernel=kernel, inclusion_value=inclusion, seed=42, **kwargs)  # type: ignore[arg-type]
-    train_seconds = time.monotonic() - t_train
-
-    step = _StepModel(
-        predict=clf.predict_proba,
-        torch_model=None,
-        backend=clf.backend,
-        device="cuda" if clf.backend == "cuml" else "cpu",
-    )
-    return step, threshold, n_labels, {"train_seconds": train_seconds, "xcal_seconds": xcal_seconds}, {}
-
-
-# ------------------------------------------------------------------
 # Supervised skyline (issue #3322)
 # ------------------------------------------------------------------
 
@@ -2947,7 +2143,7 @@ def _skyline_fit_and_score(
     inclusion: int,
     calibrate_count: int,
     calibration_fraction: float,
-) -> tuple[dict[int, float], _StepModel, dict[str, float], float]:
+) -> tuple[dict[int, float], StepModel, dict[str, float], float]:
     """Train one fully-supervised head and score *score_ids* with it.
 
     The whole point of the skyline is that it differs from a mortal step in
@@ -3050,7 +2246,7 @@ def _skyline_arm_rows(
     )
     wf, wn = cm.inclusion_weights(inclusion)
 
-    def _row(name: str, score_map: dict[int, float], step: _StepModel, timings: dict[str, float], secs: float):
+    def _row(name: str, score_map: dict[int, float], step: StepModel, timings: dict[str, float], secs: float):
         scores = np.array([score_map[cid] for cid in ordered_test], dtype=np.float64)
         o_thr, _o_cost, _o_fpr, _o_fnr = cm.oracle_cut(scores, test_labels, wf, wn)
         if not np.isfinite(o_thr):
@@ -3145,7 +2341,7 @@ def _skyline_xfit_scores(
     calibrate_count: int,
     calibration_fraction: float,
     seed: int,
-) -> tuple[dict[int, float], _StepModel, dict[str, float], float] | None:
+) -> tuple[dict[int, float], StepModel, dict[str, float], float] | None:
     """Cross-fitted test-side skyline scores: every item scored by a head that never saw it.
 
     Partitions *ordered_test* into :data:`~vtscore.eval.transfer_rules.HONEST_ORACLE_FOLDS`
@@ -3172,7 +2368,7 @@ def _skyline_xfit_scores(
     scores: dict[int, float] = {}
     timings = {"train_seconds": 0.0, "xcal_seconds": 0.0}
     total_secs = 0.0
-    step: _StepModel | None = None
+    step: StepModel | None = None
     for part in np.array_split(order, HONEST_ORACLE_FOLDS):
         held = [ordered_test[i] for i in part]
         if not held:
@@ -3777,7 +2973,7 @@ def simulate_voting_iterations(  # noqa: C901
     # assembled.
     region_aware = any(clips_dict[cid].get("patch_grid") is not None for cid in clips_dict)
 
-    # `region_voting` is a request, not a guarantee: `_good_training_vec` pools
+    # `region_voting` is a request, not a guarantee: `good_training_vec` pools
     # the ground-truth box only when the media carries a stored `patch_grid`,
     # and falls back to the whole-image embedding otherwise - which is the same
     # condition `region_aware` above tests.  On a single-vector embedder that
@@ -3909,7 +3105,7 @@ def simulate_voting_iterations(  # noqa: C901
     # examples from the positives here when no text sort is available.  Cheap to
     # build once up front.
     pool_labels = {cid: (1.0 if media_is_positive(clips_dict[cid], target_category) else 0.0) for cid in sim_ids}
-    step: _StepModel | None = None
+    step: StepModel | None = None
     threshold = 0.5
     #: The selector's threshold - cut ``acq_inclusion_offset`` steps below the
     #: reporting one.  Kept as its own name so the two jobs cannot silently
